chain/ethereum: Cache failed call results
We want to cache failed call results in the in-memory eth call cache for
the duration of a block so that failed declared calls do not need to be
repeated when the mapping calls `ethereum.call`.

We still do not cache failed results in the store as the store ignores
CallResult::Null

This also uses CallResult::Null instead of ContractCallError::Revert
internally. All we ever did with that error was log it and return `None`
from RuntimeAdapter::eth_call. With these changes, we just log a bit
earlier and return the same result as before