chain/ethereum, graph, store: Avoid double-counting declared calls
With declared calls, we seemingly make an eth call at least twice: first in
the `DecoderHook` when we either get the return value from the store or from
RPC, and then a second time when the mapping uses `ethereum.call` to access
the return value.

In both cases, we observed the call in the host metrics, inflating the call
count. We now annotate the return value with a `CallSource` that tells us
where we found the return value and do not mark that call in the metrics if
we just looked up a declared call that we made from the `DecoderHook`.